Alle Komponenten ausgelagert,, Problem: anpassen der Farben auf skeletonUI
