+++ src/components/index.ts
@@ -1,1 +1,5 @@
 export { default as RechtlicheHinweise } from './about/rechtlicheHinweise.svelte';
+export { default as Produktvideo } from './about/produktvideo.svelte';
+export { default as Projektdokumentation } from './about/projektdokumentation.svelte';
+export { default as KarusellProjektbeteiligte } from './about/karusellProjektbeteiligte.svelte';
+export { default as MediaButtonLogos } from './about/mediaButtonLogos.svelte';
